// packages/core/src/models/Formidable/Field/Core/field.props.interface.ts
import { FieldType } from '@/models/Formidable/Field/Core/field.types.enum';

/**
 * Basic props for a Formidable Field
 */
export interface IFormidableFieldProps<T> {
    fieldType: FieldType;
    name?: string;
    label?: string;
    id?: string;
    required?: boolean;
    disabled?: boolean;
    value: T | null;
}